+++ Tests/BugsnagPerformanceTests/BugsnagPerformanceConfigurationTests.mm
@@ -22,8 +22,10 @@ static NSString *const performanceBundleVersion = @"PerformanceBundleVersion";
 static NSString *const performanceReleaseStage = @"PerformanceReleaseStage";
 static NSString *const performanceReleaseStage1 = @"PerformanceEnabledReleaseStage1";
 static NSString *const performanceReleaseStage2 = @"PerformanceEnabledReleaseStage2";
+static NSString *const performanceServiceName = @"PerformanceServiceName";
 static NSString *const performanceEndpoint = @"PerformanceEndpoint";
 static NSArray *const performanceEnabledReleaseStages = @[performanceReleaseStage1, performanceReleaseStage2];
+static NSArray *const performanceTracePropagationUrls = @[@"https://my.company.com/.*", @"https://somewhere.com/[0-9]+/abc/*"];
 
 static NSString *const bugsnagApiKey = @"PerfromanceApiKey";
 static NSString *const bugsnagAppVersion = @"PerfromanceAppVersion";
@@ -88,6 +90,14 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
     XCTAssertFalse([config shouldSendReports]);
 }
 
+- (void)assertConfig:(BugsnagPerformanceConfiguration *)config tracePropagationUrlsAre:(NSArray<NSString *> *) regexStrings {
+    XCTAssertEqual(config.tracePropagationUrls.count, regexStrings.count);
+    for(NSString *regexString: regexStrings) {
+        NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:regexString options:0 error:nil];
+        XCTAssertTrue([config.tracePropagationUrls containsObject:regex]);
+    }
+}
+
 - (void)testLoadConfigLoadsWhenAllValuesAreInPerformanceDictionary {
     auto config = [BugsnagPerformanceConfiguration loadConfigWithInfoDictionary:@{
         @"bugsnag": @{
@@ -97,7 +107,12 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
                 @"bundleVersion": performanceBundleVersion,
                 @"releaseStage": performanceReleaseStage,
                 @"enabledReleaseStages": performanceEnabledReleaseStages,
+                @"serviceName": performanceServiceName,
                 @"endpoint": performanceEndpoint,
+                @"attributeArrayLengthLimit": @100,
+                @"attributeStringValueLimit": @200,
+                @"attributeCountLimit": @50,
+                @"tracePropagationUrls": performanceTracePropagationUrls,
                 @"autoInstrumentAppStarts": @(NO),
                 @"autoInstrumentViewControllers": @(NO),
                 @"autoInstrumentNetworkRequests": @(YES),
@@ -112,6 +127,10 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
     XCTAssertTrue([config.enabledReleaseStages containsObject:performanceReleaseStage1]);
     XCTAssertTrue([config.enabledReleaseStages containsObject:performanceReleaseStage2]);
     XCTAssertEqualObjects([config.endpoint description], performanceEndpoint);
+    XCTAssertEqual(config.attributeArrayLengthLimit, (NSUInteger)100);
+    XCTAssertEqual(config.attributeStringValueLimit, (NSUInteger)200);
+    XCTAssertEqual(config.attributeCountLimit, (NSUInteger)50);
+    [self assertConfig:config tracePropagationUrlsAre:performanceTracePropagationUrls];
     XCTAssertFalse(config.autoInstrumentAppStarts);
     XCTAssertFalse(config.autoInstrumentViewControllers);
     XCTAssertTrue(config.autoInstrumentNetworkRequests);
@@ -131,7 +150,12 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
                 @"bundleVersion": performanceBundleVersion,
                 @"releaseStage": performanceReleaseStage,
                 @"enabledReleaseStages": performanceEnabledReleaseStages,
+                @"serviceName": performanceServiceName,
                 @"endpoint": performanceEndpoint,
+                @"attributeArrayLengthLimit": @100,
+                @"attributeStringValueLimit": @200,
+                @"attributeCountLimit": @50,
+                @"tracePropagationUrls": performanceTracePropagationUrls,
                 @"autoInstrumentAppStarts": @(NO),
                 @"autoInstrumentViewControllers": @(NO),
                 @"autoInstrumentNetworkRequests": @(YES),
@@ -145,7 +169,12 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
     XCTAssertEqual(config.enabledReleaseStages.count, performanceEnabledReleaseStages.count);
     XCTAssertTrue([config.enabledReleaseStages containsObject:performanceReleaseStage1]);
     XCTAssertTrue([config.enabledReleaseStages containsObject:performanceReleaseStage2]);
+    XCTAssertEqualObjects(config.serviceName, performanceServiceName);
     XCTAssertEqualObjects([config.endpoint description], performanceEndpoint);
+    XCTAssertEqual(config.attributeArrayLengthLimit, (NSUInteger)100);
+    XCTAssertEqual(config.attributeStringValueLimit, (NSUInteger)200);
+    XCTAssertEqual(config.attributeCountLimit, (NSUInteger)50);
+    [self assertConfig:config tracePropagationUrlsAre:performanceTracePropagationUrls];
     XCTAssertFalse(config.autoInstrumentAppStarts);
     XCTAssertFalse(config.autoInstrumentViewControllers);
     XCTAssertTrue(config.autoInstrumentNetworkRequests);
@@ -183,6 +212,34 @@ static NSArray *const bugsnagEnabledReleaseStages = @[bugsnagReleaseStage1, bugs
 - (void)testShouldSetIncludeApiKeyInTheDefaultEndpoint {
     auto config = [[BugsnagPerformanceConfiguration alloc] initWithApiKey:@"0123456789abcdef0123456789abcdef"];
     XCTAssertEqualObjects(config.endpoint.absoluteString, @"https://0123456789abcdef0123456789abcdef.otlp.bugsnag.com/v1/traces");
+}
+
+- (void)testLimits {
+#define MIN_ATTRIBUTE_COUNT_LIMIT (uint64_t)1
+#define MAX_ATTRIBUTE_COUNT_LIMIT (uint64_t)500
+#define DEFAULT_ATTRIBUTE_COUNT_LIMIT (uint64_t)100
+
+#define MIN_ATTRIBUTE_ARRAY_LENGTH_LIMIT (uint64_t)1
+#define MAX_ATTRIBUTE_ARRAY_LENGTH_LIMIT (uint64_t)10000
+#define DEFAULT_ATTRIBUTE_ARRAY_LENGTH_LIMIT (uint64_t)1000
+
+#define MIN_ATTRIBUTE_STRING_VALUE_LIMIT (uint64_t)1
+#define MAX_ATTRIBUTE_STRING_VALUE_LIMIT (uint64_t)10000
+#define DEFAULT_ATTRIBUTE_STRING_VALUE_LIMIT (uint64_t)1024
+
+    auto config = [[BugsnagPerformanceConfiguration alloc] initWithApiKey:@"0123456789abcdef0123456789abcdef"];
+    XCTAssertEqual(config.attributeArrayLengthLimit, DEFAULT_ATTRIBUTE_ARRAY_LENGTH_LIMIT);
+    XCTAssertEqual(config.attributeStringValueLimit, DEFAULT_ATTRIBUTE_STRING_VALUE_LIMIT);
+
+    config.attributeArrayLengthLimit = MAX_ATTRIBUTE_ARRAY_LENGTH_LIMIT + 1;
+    config.attributeStringValueLimit = MAX_ATTRIBUTE_STRING_VALUE_LIMIT + 1;
+    XCTAssertEqual(config.attributeArrayLengthLimit, MAX_ATTRIBUTE_ARRAY_LENGTH_LIMIT);
+    XCTAssertEqual(config.attributeStringValueLimit, MAX_ATTRIBUTE_STRING_VALUE_LIMIT);
+
+    config.attributeArrayLengthLimit = 0;
+    config.attributeStringValueLimit = 0;
+    XCTAssertEqual(config.attributeArrayLengthLimit, DEFAULT_ATTRIBUTE_ARRAY_LENGTH_LIMIT);
+    XCTAssertEqual(config.attributeStringValueLimit, DEFAULT_ATTRIBUTE_STRING_VALUE_LIMIT);
 }
 
 @end